Front end

+++ services/task-service/controllers/tasks.js
@@ -5,13 +5,12 @@ const mongoose = require('mongoose')
 
 function allocateTask(users,taskId){
   try{
-    console.log(users)
     users.forEach(userToAdd=>{
     const newUserTask = new UserTask({
-      userId:userToAdd,
+      userId:userToAdd.userId,
+      username:userToAdd.username,
       taskId
     })
-    console.log(newUserTask)
     newUserTask.save()
   })}
   catch(err){
@@ -24,7 +23,7 @@ function allocateTask(users,taskId){
 }
 function deallocateTask(users,taskId){
   try{[].forEach(userToRemove=>{
-    UserTask.findOneAndRemove({userId:userToRemove,taskId})
+    UserTask.findOneAndRemove({userId:userToRemove.userId,taskId})
     })
   }
   catch(err){
@@ -41,9 +40,7 @@ exports.createSubTask = async(req,res)=>{
   const masterTaskId = req.params.taskId
   const {name,startDate,endDate} = req.body
   const master =  await Task.findById(masterTaskId)
-  const users = []
-  if(req.body.users!=undefined)
-  users.push(req.body.users)
+  const users = req.body.users
   const newTask = new Task({
     name,
     startDate,
@@ -54,6 +51,7 @@ exports.createSubTask = async(req,res)=>{
   newTask
   .save()
   .then(async () => {
+    if(users!=undefined)
   allocateTask(users,newTask._id)
   return res.status(200).json({
       success: true,
@@ -72,20 +70,24 @@ exports.createSubTask = async(req,res)=>{
 
 exports.createTask = async (req,res)=>{
   const projectId = req.params.projectId
-      const {name,startDate,endDate} = req.body
-      const users = []
-      if(req.body.users!=undefined)
-      users.push(req.body.users)
+      const {name,startDate,endDate,priority,description,masterTaskId,depedencies,toBeApproved} = req.body
+      const users = req.body.usernames
       const newTask = new Task({
         name,
         startDate,
         endDate,
+        priority,
+        description,
+        masterTaskId,
+        toBeApproved,
+        depedencies,
         projectId
       })
       newTask
       .save()
       .then(async () => {
-      allocateTask(users,newTask._id)
+        if(users!=undefined)
+        allocateTask(users,newTask._id)
       return res.status(200).json({
           success: true,
           message: 'Task is created',
@@ -100,6 +102,47 @@ exports.createTask = async (req,res)=>{
         });
       });
 }
+exports.getTask = async (req, res) => {
+  const taskId = req.params.taskId;
+  const task = await Task.findById(taskId);
+  const userTask = await UserTask.findOne({ taskId, userId: req.user._id });
+
+  if (userTask) {
+    userTask.lastViewed = Date.now();
+    await userTask.save();
+  }
+
+  const startDate = task.startDate.toLocaleDateString('en-US', { month: 'long', day: 'numeric' });
+
+  return res.status(200).json({ ...task.toObject(), startDate });
+};
+
+exports.getRecentTask = async (req, res) => {
+  const recentTasks = await UserTask.find({ userId: req.user._id })
+    .sort({ lastViewed: "desc" })
+    .limit(3);
+
+  // Get an array of task IDs from the recentTasks array
+  const taskIds = recentTasks.map(({ taskId }) => taskId);
+
+  // Get an array of tasks with their names and project IDs
+  const tasks = await Task.find({ _id: { $in: taskIds } }, { name: 1, projectId: 1, createdAt: 1 });
+
+  // Map the tasks array to include the lastViewed date for each task
+  const tasksWithLastViewed = tasks.map((task) => {
+    const recentTask = recentTasks.find(({ taskId }) => taskId.match(task._id));
+
+    if (recentTask.lastViewed == undefined) {
+      const lastViewed = new Date(recentTask.lastViewed);
+      return { ...task._doc, lastViewed: lastViewed.toLocaleString("default", { month: "short", day: "numeric" }) };
+    } else {
+      const createdAt = new Date(task.createdAt);
+      return { ...task._doc, lastViewed: createdAt.toLocaleString("default", { month: "short", day: "numeric" }) };
+    }
+  });
+
+  return res.status(200).json(tasksWithLastViewed );
+};
 
 exports.updateTask = async (req, res)=> {
   const id = req.params.taskId;
@@ -114,7 +157,7 @@ exports.updateTask = async (req, res)=> {
   await Task.update({ _id:id }, { $set:updateObject })
     .exec()
     .then(() => {
-      res.status(200).json({
+      res.status(200).json({ 
         success: true,
         message: 'Task is updated',
         updateProject: updateObject,
@@ -153,21 +196,27 @@ exports.getByProjectId = async (req, res) => {
 
   try {
     // Fetch all the master tasks for the project
-    const masterTasks = await Task.find({ projectId: projectId, masterTaskId: null });
+    const masterTasks = await Task.find({ projectId, masterTaskId: { $in: ["", null] } });
 
     // Fetch all the subtasks for each master task and group them under their respective master task
     const tasks = await Promise.all(
       masterTasks.map(async (masterTask) => {
         const subTasks = await Task.find({ projectId: projectId, masterTaskId: masterTask._id });
-        return { ...masterTask._doc, subTasks };
+        const userTasks = await UserTask.find({ taskId: masterTask._id}||{taskId: subTask._id} ,{userId:1});
+        
+        const assignedTo = userTasks.filter(user => user != null).map(user => user.userId); // extract username property and return as an array of strings
+        
+
+        // Format the startDate and endDate fields
+        const options = { month: 'short', day: 'numeric' };
+        const startDate = new Date(masterTask.startDate).toLocaleString('en-US', options);
+        const endDate = new Date(masterTask.endDate).toLocaleString('en-US', options);
+
+        return { ...masterTask._doc, subTasks, assignedTo, startDate, endDate };
       })
     );
-
-    return res.status(200).json({
-      success: true,
-      message: 'A list of all tasks',
-      Cause: tasks,
-    });
+    console.log(tasks)
+    return res.status(200).json(tasks);
   } catch (err) {
     return res.status(500).json({
       success: false,
@@ -179,13 +228,20 @@ exports.getByProjectId = async (req, res) => {
 
 exports.getByUserId = async (req, res) => {
   const id = req.user._id;
+  
   const allTasks = [];
   try {
     const allTaskRecords = await UserTask.find({ userId: id });
+    
     for (const taskRecord of allTaskRecords) {
+      
       const task = await Task.findById(taskRecord.taskId);
-      allTasks.push(task);
+      
+      const startDate = task.startDate.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })
+      const endDate = task.endDate.toLocaleString('en-US', { month: 'short', day: 'numeric' })
+      allTasks.push({ ...task._doc, startDate,endDate });
     }
+    
     res.status(200).json(allTasks);
   } catch (err) {
     res.status(500).json({
@@ -195,3 +251,25 @@ exports.getByUserId = async (req, res) => {
     });
   }
 };
+
+exports.completeTasks = async (req, res) => {
+  try {
+    const task = await Task.findById(req.params.taskId)
+    await Task.findByIdAndUpdate(req.params.taskId, { $set: { isCompleted:(!task.isCompleted) } }).exec()
+    res.status(200).json('Task updated successfully');
+  } catch (err) {
+    console.error(err);
+    res.status(500).json('Error updating task');
+  }
+}
+
+exports.approveTasks = async (req, res) => {
+  try {
+    const task = await Task.findById(req.params.taskId)
+    await Task.findByIdAndUpdate(req.params.taskId, { $set: { isApproved:(!task.isApproved) } }).exec()
+    res.status(200).json('Task updated successfully');
+  } catch (err) {
+    console.error(err);
+    res.status(500).json('Error updating task');
+  }
+}
